requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blis==0.7.11
cachetools==6.2.1
catalogue==2.0.10
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz#sha256=1075c2aa2bc2fee105ab6e90a01a5d1a428c9f5b20a1fa003dc2cb6a438d295e
exceptiongroup==1.3.0
google-auth==2.42.1
google-genai==1.47.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.3.1
MarkupSafe==3.0.3
murmurhash==1.0.13
numpy==1.26.4
packaging==25.0
pathlib_abc==0.1.1
pathy==0.11.0
preshed==3.0.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
requests==2.32.5
rsa==4.9.1
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
tenacity==9.1.2
thinc==8.2.5
tqdm==4.67.1
typer==0.9.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wasabi==1.1.3
weasel==0.3.4
websockets==15.0.1

fastapi
uvicorn
gradio
torch
pyyaml
pytest
python-dotenv
transformers
scipy